# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-PERSON ",
        "description": "Documentação da API-PERSON com Swagger Service.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/unidades": {
            "get": {
                "tags": [
                    "Unidade"
                ],
                "summary": "Lista todos as unidades",
                "description": "Retorna uma lista de unidades armazenados no banco de dados.",
                "operationId": "ac4d45815202ccec305966db0d0b3fcc",
                "responses": {
                    "200": {
                        "description": "Lista de unidades retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Unidade"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum unidades encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não há unidades cadastrados!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao buscar os unidades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar os unidades."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Unidade"
                ],
                "summary": "Cria uma nova unidade",
                "description": "Registra uma nova unidade no banco de dados.",
                "operationId": "30456808cb280227b9333f98ee71c6b4",
                "requestBody": {
                    "description": "Dados necessários para criar uma nova unidade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_nome",
                                    "pes_data_nascimento",
                                    "pes_sexo",
                                    "pes_mae",
                                    "pes_pai"
                                ],
                                "properties": {
                                    "pes_nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "pes_data_nascimento": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "1978-08-23"
                                    },
                                    "pes_sexo": {
                                        "type": "string",
                                        "example": "Masculino"
                                    },
                                    "pes_mae": {
                                        "type": "string",
                                        "example": "Maria Aparecida da Silva"
                                    },
                                    "pes_pai": {
                                        "type": "string",
                                        "example": "Cícero Joaquim da Silva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Unidade criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade criado com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Unidade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o unidade",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o unidade."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Unidade": {
                "title": "Unidade",
                "description": "Modelo da Unidade",
                "required": [
                    "unid_nome",
                    "unid_sigla"
                ],
                "properties": {
                    "unid_id": {
                        "description": "ID da Unidade",
                        "type": "integer",
                        "example": 1
                    },
                    "unid_nome": {
                        "description": "Nome da unidade",
                        "type": "string",
                        "example": "Secretária de Planejamento"
                    },
                    "unid_sigla": {
                        "description": "Sigla da unidade",
                        "type": "string",
                        "example": "SEPLAG"
                    }
                },
                "type": "object"
            }
        }
    },
...
}